requirements.txt
# Flask core
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Admin>=1.6.1

# Database
SQLAlchemy>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Security
Werkzeug>=3.0.0

# Scheduling
APScheduler>=3.10.0

# PDF Processing
docling>=2.5.0

# OCR Support (via docling extras)
easyocr>=1.7.0

# Hugging Face Hub (with optional XET support for faster downloads)
huggingface_hub[hf_xet]>=0.20.0

# LLM - Google Gemini
google-genai>=0.3.0

# Production server
gunicorn>=21.0.0

# Utilities
python-dateutil>=2.8.0

# PyTorch (for GPU support - installed separately in Docker)
# torch>=2.1.0
# torchvision>=0.16.0
